create print functions
