+++ test/print.js
@@ -1,22 +1,6 @@
-const pluralize = require('pluralize')
+const { stringFileTests, stringAllInfo } = require('../src/print')
 const test = require('ava')
-const { formatTestList } = require('find-test-names')
 const input = require('./tagged.json')
-
-/**
- * Outputs a string representation of the json test results object,
- * like a tree of suites and tests.
- */
-function stringFileTests(fileName, fileInfo) {
-  const testCount = pluralize('test', fileInfo.counts.tests, true)
-  const headerLine = fileInfo.counts.pending
-    ? `${fileName} (${testCount} ${fileInfo.counts.pending} pending)`
-    : `${fileName} (${testCount})`
-
-  const body = formatTestList(fileInfo.tests)
-
-  return headerLine + '\n' + body + '\n'
-}
 
 test('prints tests from one file', (t) => {
   t.plan(2)
@@ -40,6 +24,17 @@ test('prints tests with inner suites', (t) => {
 
   // does not change the input
   t.deepEqual(fileInfo, copy)
+  // console.log(str)
+  t.snapshot(str)
+})
+
+test('prints all tests', (t) => {
+  t.plan(2)
+  const copy = JSON.parse(JSON.stringify(input))
+  const str = stringAllInfo(copy)
+
+  // does not change the input
+  t.deepEqual(input, copy)
   // console.log(str)
   t.snapshot(str)
 })
