+++ test/snapshots/print.js.md
@@ -23,3 +23,18 @@ Generated by [AVA](https://avajs.dev).
       └─ inner suite␊
         └─ shows something! [@user]␊
     `
+
+## prints all tests
+
+> Snapshot 1
+
+    `cypress/e2e/spec.js (2 tests)␊
+    └─ parent suite [@main]␊
+      ├─ works well enough␊
+      └─ inner suite␊
+        └─ shows something! [@user]␊
+    ␊
+    cypress/e2e/featureA/user.js (2 tests 1 pending)␊
+    ├─ works [@user]␊
+    └⊙ needs to be written [@alpha]␊
+    `
